# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "应用系统接口基于Golang Gin框架开发，Gin是一个非常受欢迎的Golang Web框架，在GitHub上已经有47k的星星，后续开发可以参考官方手册,https://gin-gonic.com/",
        "title": "应用系统接口服务",
        "termsOfService": "http://127.0.0.1:8080/api/v1/token",
        "contact": {
            "name": "Cameron",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "http://127.0.0.1:8080",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/information/sample/delete": {
            "post": {
                "description": "删除样本数据",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "标签"
                ],
                "summary": "删除样本列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "默认为 空",
                        "name": "id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "默认为 空",
                        "name": "token",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ExecSuccess"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ServerError"
                        }
                    }
                }
            }
        },
        "/api/v1/information/sample/get": {
            "get": {
                "description": "获取样本列表数据",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "标签"
                ],
                "summary": "获取样本列表数据",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "默认为 1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 25,
                        "description": "默认为 25",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "默认为 空",
                        "name": "level",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "默认为 空",
                        "name": "token",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "默认为 空",
                        "name": "md5",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "默认为 空",
                        "name": "sha1",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.JsonOut"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.SampleInfo"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ServerError"
                        }
                    }
                }
            }
        },
        "/api/v1/information/sample/insert": {
            "post": {
                "description": "新增样本数据",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "标签"
                ],
                "summary": "新增样本信息",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 70,
                        "description": "默认为 空",
                        "name": "level",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "2a57220fe8f64481b1311c892b788da5",
                        "description": "默认为 空",
                        "name": "md5",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "ff4cd7d8ee07f35037e834cc0f356f5fa159c871",
                        "description": "默认为 空",
                        "name": "sha1",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "admin",
                        "description": "默认为 空",
                        "name": "operator",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "默认为 空",
                        "name": "token",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ExecSuccess"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ServerError"
                        }
                    }
                }
            }
        },
        "/api/v1/information/sample/update": {
            "post": {
                "description": "更新样本数据",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "标签"
                ],
                "summary": "更新样本信息",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "默认为 空",
                        "name": "level",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "默认为 空",
                        "name": "id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "默认为 空",
                        "name": "token",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ExecSuccess"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ServerError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.ExecSuccess": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "type": "boolean",
                    "example": true
                },
                "message": {
                    "type": "string",
                    "example": "OK"
                }
            }
        },
        "model.JsonOut": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "OK"
                }
            }
        },
        "model.SampleInfo": {
            "type": "object",
            "properties": {
                "create_time": {
                    "type": "string",
                    "example": "2022-10-20 11:00:01"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "level": {
                    "type": "integer",
                    "example": 70
                },
                "md5": {
                    "type": "string",
                    "example": "2a57220fe8f64481b1311c892b788da5"
                },
                "operator": {
                    "type": "string",
                    "example": "admin"
                },
                "sha1": {
                    "type": "string",
                    "example": "ff4cd7d8ee07f35037e834cc0f356f5fa159c871"
                }
            }
        },
        "model.ServerError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "data": {
                    "type": "boolean",
                    "example": false
                },
                "message": {
                    "type": "string",
                    "example": "服务器内部错误"
                }
            }
        }
    }
}